[net] bonding: move bond_attach/detach_slave in the proper position
Message-id: <1416574342-24493-61-git-send-email-nikolay@redhat.com>
Patchwork-id: 100468
O-Subject: [RHEL7.2 PATCH BZ 1166647 060/336] bonding: RHEL-specific: move bond_attach/detach_slave in the proper position
Bugzilla: 1166647
RH-Acked-by: Neil Horman <nhorman@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

RHEL specific.

This should move the attach/detach slave in the end of bond_enslave and
the beginning of bond_release_one respectively since that is the where
the slaves get linked in upstream (using neighbor lists though).
This will help us apply a couple of future fixes and patches that depend
on this enslave/release organization.

Signed-off-by: Nikolay Aleksandrov <nikolay@redhat.com>

Signed-off-by: Jiri Benc <jbenc@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>