[fs] vfs: export check_sticky()
Message-id: <20141125163449.18115.69997.stgit@warthog.procyon.org.uk>
Patchwork-id: 100926
O-Subject: [PATCH 26/47] vfs: export check_sticky()
Bugzilla: 985875
RH-Acked-by: Steve Best <sbest@redhat.com>

From: Miklos Szeredi <mszeredi@suse.cz>

It's already duplicated in btrfs and about to be used in overlayfs too.

Move the sticky bit check to an inline helper and call the out-of-line
helper only in the unlikly case of the sticky bit being set.

Signed-off-by: Miklos Szeredi <mszeredi@suse.cz>
Original-commit: cbdf35bcb833bfd00f0925d7a9a33a21f41ea582
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>