[fs] ovl: fix race in private xattr checks
Message-id: <20141127114649.9170.92959.stgit@warthog.procyon.org.uk>
Patchwork-id: 100986
O-Subject: [RHEL7 PATCH 3/7] ovl: fix race in private xattr checks
Bugzilla: 985875
RH-Acked-by: Alexander Viro <aviro@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

From: Miklos Szeredi <mszeredi@suse.cz>

Xattr operations can race with copy up.  This does not matter as long as
we consistently fiter out "trunsted.overlay.opaque" attribute on upper
directories.

Previously we checked parent against OVL_PATH_MERGE.  This is too general,
and prone to race with copy-up.  I.e. we found the parent to be on the
lower layer but ovl_dentry_real() would return the copied-up dentry,
possibly with the "opaque" attribute.

So instead use ovl_path_real() and decide to filter the attributes based on
the actual type of the dentry we'll use.

Signed-off-by: Miklos Szeredi <mszeredi@suse.cz>
Original-commit: 521484639ec19a6f1ed56de6993feb255f5f676c
---

Signed-off-by: Jarod Wilson <jarod@redhat.com>