[tools] perf: Fix a race condition in perf_remove_from_context()
Message-id: <1418163608-7772-5-git-send-email-jolsa@redhat.com>
Patchwork-id: 101215
O-Subject: [PATCH RHEL7.1 BZ1169436 004/359] perf: Fix a race condition in perf_remove_from_context()
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit 3577af70a2ce4853d58e57d832e687d739281479
Author: Cong Wang <cwang@twopensource.com>
Date: Tue Sep 2 15:27:20 2014 -0700

description
===========
We saw a kernel soft lockup in perf_remove_from_context(),
it looks like the `perf` process, when exiting, could not go
out of the retry loop. Meanwhile, the target process was forking
a child. So either the target process should execute the smp
function call to deactive the event (if it was running) or it should
do a context switch which deactives the event.

It seems we optimize out a context switch in perf_event_context_sched_out(),
and what's more important, we still test an obsolete task pointer when
retrying, so no one actually would deactive that event in this situation.
Fix it directly by reloading the task pointer in perf_remove_from_context().

This should cure the above soft lockup.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>