[tools] perf/x86: Fix pp without LBR
Message-id: <1418163608-7772-13-git-send-email-jolsa@redhat.com>
Patchwork-id: 101223
O-Subject: [PATCH RHEL7.1 BZ1169436 012/359] perf/x86: Fix :pp without LBR
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit 03de874aa76ac0adcf6f56ebf3de623d09a5dde3
Author: Andi Kleen <ak@linux.intel.com>
Date: Thu Aug 7 17:08:54 2014 -0700

description
===========
This fixes a side effect of Kan's earlier patch to probe the LBRs at boot
time. Normally when the LBRs are disabled cycles:pp is disabled too.
So for example cycles:pp doesn't work.

However this is not needed with PEBSv2 and later (Haswell) because
it does not need LBRs to correct the IP-off-by-one.

So add an extra check for PEBSv2 that also allows :pp
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>