[tools] perf/x86/intel: Mark MEM_LOAD_UOPS_MISS_RETIRED as precise on SNB
Message-id: <1418163608-7772-14-git-send-email-jolsa@redhat.com>
Patchwork-id: 101224
O-Subject: [PATCH RHEL7.1 BZ1169436 013/359] perf/x86/intel: Mark MEM_LOAD_UOPS_MISS_RETIRED as precise on SNB
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit 9d8e3f9693245415db0b7c58551a91fa9fd1f9c7
Author: Stephane Eranian <eranian@google.com>
Date: Fri Sep 13 13:16:46 2013 -0700

description
===========
On Intel SNB (SNB, SNB-EP), the event MEM_LOAD_UOPS_MISS_RETIRED
supports PEBS. It was missing for the SNB PEBS event constraint
table thereby preventing any measurement with PEBS for it.

This patch adds the event to the PEBS table for SNB.

WARNING: it should be noted that this event like a few others
are subject to the erratum BT241 for Xeon E5 (SNB-EP). As such,
the event may undercount when used with PEBS unless the
workaround is implemented. But without this patch and just the
workaround, the kernel would not allow precise sampling on this
event. BT241 is documented in:

  http://www.intel.com/content/dam/www/public/us/en/documents/specification-updates/xeon-e5-family-spec-update.pdf
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>