[tools] Revert: perf: Do not allow optimized switch for non-cloned events
Message-id: <1418163608-7772-28-git-send-email-jolsa@redhat.com>
Patchwork-id: 101238
O-Subject: [PATCH RHEL7.1 BZ1169436 027/359] Revert "perf: Do not allow optimized switch for non-cloned events"
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit 802c8a61d4c9c794db863dcabb0006ab001a651b
Author: Jiri Olsa <jolsa@redhat.com>
Date: Fri Sep 12 13:18:28 2014 +0200

description
===========
This reverts commit 1f9a7268c67f0290837aada443d28fd953ddca90.

With the fix of the initial state for the cloned event we now correctly
handle the error described in:

  1f9a7268c67f perf: Do not allow optimized switch for non-cloned events

so we can revert it.

I made an automated test for this, but its not suitable for automated
perf tests framework. It needs to be customized for each machine (the
more cpu the higher numbers for GROUPS/WORKERS/BYTES) and it could take
longer time to hit the issue.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>