[tools] perf/x86: Tone down kernel messages when the PMU check fails in a virtual environment
Message-id: <1418163608-7772-34-git-send-email-jolsa@redhat.com>
Patchwork-id: 101244
O-Subject: [PATCH RHEL7.1 BZ1169436 033/359] perf/x86: Tone down kernel messages when the PMU check fails in a virtual environment
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit cc6cd47e7395bc05c5077009808b820633eb3f18
Author: Wei Huang <wei@redhat.com>
Date: Wed Sep 24 22:55:14 2014 -0500

description
===========
PMU checking can fail due to various reasons. On native machine, this
is mostly caused by faulty hardware and it is reasonable to use
KERN_ERR in reporting. However, when kernel is running on virtualized
environment, this checking can fail if virtual PMU is not supported
(e.g. KVM on AMD host). It is annoying to see an error message on
splash screen, even though we know such failure is benign on
virtualized environment.

This patch checks if the kernel is running in a virtualized environment.
If so, it will use KERN_INFO in reporting, which reduces the syslog
priority of them. This patch was tested successfully on KVM.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>