[tools] perf/hists/browser: Override ui_browser refresh_dimensions method
Message-id: <1418163608-7772-72-git-send-email-jolsa@redhat.com>
Patchwork-id: 101282
O-Subject: [PATCH RHEL7.1 BZ1169436 071/359] perf hists browser: Override ui_browser refresh_dimensions method
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit 357cfff1c2173b7aedc0e521a2146b78396171e0
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Jul 1 17:01:01 2014 -0300

description
===========
This requires some more work so that we can really just use the width of
current entries when we want to partition the screen.

Right now its just a prep patch so that we can have where to update
ui_browser->rows when introducing the column headers line, that will be
togglable, so we need to update it everytime we refresh the dimensions
of the browser.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>