[tools] perf/script: Provide additional sample information on generic events
Message-id: <1418163608-7772-94-git-send-email-jolsa@redhat.com>
Patchwork-id: 101304
O-Subject: [PATCH RHEL7.1 BZ1169436 093/359] perf script: Provide additional sample information on generic events
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit 57608cfd8827a74237d264a197722e2c99f72da4
Author: Joseph Schuchart <joseph.schuchart@tu-dresden.de>
Date: Thu Jul 10 13:50:56 2014 +0200

description
===========
To python scripts, including pid, tid, and cpu for which the event was
recorded.

At the moment, the pointer to the sample struct is passed to scripts,
which seems to be of little use.

The patch puts this information in dictionaries for easy access by
Python scripts.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>