[tools] perf/buildid-cache: Apply force option to copying kcore
Message-id: <1418163608-7772-102-git-send-email-jolsa@redhat.com>
Patchwork-id: 101312
O-Subject: [PATCH RHEL7.1 BZ1169436 101/359] perf buildid-cache: Apply force option to copying kcore
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit 5173fbb8a11b2857aeec9e5f4e9568d4e1b84dbd
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Mon Jul 14 13:02:38 2014 +0300

description
===========
Currently a copy of kcore is not made if there is one already with the
same modules at the same addresses.

Change this to make a copy anyway if the force (-f) option is also used.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>