[tools] perf/symbols: Record whether a dso is 64-bit
Message-id: <1418163608-7772-103-git-send-email-jolsa@redhat.com>
Patchwork-id: 101313
O-Subject: [PATCH RHEL7.1 BZ1169436 102/359] perf symbols: Record whether a dso is 64-bit
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit c6d8f2a4a0c5e366330a6a2a94c06b652f4ca554
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Mon Jul 14 13:02:41 2014 +0300

description
===========
Add a flag to 'struct dso' to record if the dso is 64-bit or not.
Update the flag when reading the ELF.

This is needed for instruction decoding.  For example, x86 instruction
decoding depends on whether or not the 64-bit instruction set is used.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>