[tools] perf/session: Flag if the event stream is entirely in memory
Message-id: <1418163608-7772-106-git-send-email-jolsa@redhat.com>
Patchwork-id: 101315
O-Subject: [PATCH RHEL7.1 BZ1169436 105/359] perf session: Flag if the event stream is entirely in memory
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit 919d86d3a3109d7d4f0d7347f34711ee2f8e6609
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Mon Jul 14 13:02:51 2014 +0300

description
===========
Flag if the event stream is a file that has been mmapped in one go.

This is useful, for example, if a tool needs to keep an event for later
reference.  If the new flag is set, a pointer to the event can be
retained, otherwise the event must be copied.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>