[tools] perf/thread: Allow deletion of a thread with no map groups
Message-id: <1418163608-7772-113-git-send-email-jolsa@redhat.com>
Patchwork-id: 101323
O-Subject: [PATCH RHEL7.1 BZ1169436 112/359] perf thread: Allow deletion of a thread with no map groups
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit 9608b84e4dd95341b88cad646b114811f5bccbba
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Wed Jul 16 10:19:43 2014 +0300

description
===========
It needs to be possible to call thread__delete() on a thread with no map
groups.

This is needed for a subsequent patch which deletes a thread on the
error path before map groups have been attached.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>