[tools] perf/script: Improve srcline display for BTS
Message-id: <1418163608-7772-125-git-send-email-jolsa@redhat.com>
Patchwork-id: 101335
O-Subject: [PATCH RHEL7.1 BZ1169436 124/359] perf script: Improve srcline display for BTS
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit 8066be5fe755e95ee6a858bbcf4e6b9e933e1866
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Tue Jul 22 16:17:15 2014 +0300

description
===========
Change the order of the output to put the srcline last.

It puts the branch 'from address' and 'to address' on the same line,
which is how it would be without the source line reference.

So it makes it consistent and much easier to read.

E.g. old format:

          4028fc main+0x2c (/bin/ls)
  /build/buildd/coreutils-8.20/src/ls.c:1269 =>           40d8a0 set_program_name+0x0 (/bin/ls)

new format:

          4028fc main+0x2c (/bin/ls) =>           40d8a0 set_program_name+0x0 (/bin/ls)
  /build/buildd/coreutils-8.20/src/ls.c:1269
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>