[tools] perf: Add dsos__hit_all()
Message-id: <1418163608-7772-129-git-send-email-jolsa@redhat.com>
Patchwork-id: 101339
O-Subject: [PATCH RHEL7.1 BZ1169436 128/359] perf tools: Add dsos__hit_all()
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit 1f625b0b3dd641b74ba21640a1fea19a3dd893a2
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Tue Jul 22 16:17:23 2014 +0300

description
===========
Add ability to mark all dsos as hit.

This is needed in the case of Instruction Tracing.  It takes so long to
decode an Instruction Trace that it is not worth doing just to determine
which dsos are hit.  A later patch takes this into use.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>