[tools] perf/machine: Add ability to record the current tid for each cpu
Message-id: <1418163608-7772-131-git-send-email-jolsa@redhat.com>
Patchwork-id: 101341
O-Subject: [PATCH RHEL7.1 BZ1169436 130/359] perf machine: Add ability to record the current tid for each cpu
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit b9d266baac0429f70df3f9cf751b045730d612e3
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Tue Jul 22 16:17:25 2014 +0300

description
===========
Add an array to struct machine to store the current tid running on each
cpu.

Add machine functions to get / set the tid for a cpu.

This will be used to determine the tid when decoding a per-cpu
Instruction Trace.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>