[tools] perf/machine: Fix the lifetime of the VDSO temporary file
Message-id: <1418163608-7772-138-git-send-email-jolsa@redhat.com>
Patchwork-id: 101348
O-Subject: [PATCH RHEL7.1 BZ1169436 137/359] perf machine: Fix the lifetime of the VDSO temporary file
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit d027b64001b21328cc92d35c6444e1a7a926ea76
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Wed Jul 23 14:23:00 2014 +0300

description
===========
The VDSO temporary file is unlinked when a session is deleted.  That
precludes the possibilities that there is no session or there is more
than one session.

Correctly the vdso belongs to the machine so put the information on
'struct machine' and get rid of the global variables.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>