[tools] perf: Flush ordered events in case of allocation failure
Message-id: <1418163608-7772-164-git-send-email-jolsa@redhat.com>
Patchwork-id: 101360
O-Subject: [PATCH RHEL7.1 BZ1169436 163/359] perf tools: Flush ordered events in case of allocation failure
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit d40b4a15ab2bfcfa7d946b69ca1f12c93b22d169
Author: Jiri Olsa <jolsa@kernel.org>
Date: Fri Aug 1 13:01:04 2014 -0300

description
===========
In previous patches we added a limit for ordered events queue allocation
size. If we reach this size we need to flush (part of) the queue to get
some free buffers.

The current functionality is not affected, because the limit is hard
coded to (u64) -1. The configuration code for size will come in
following patches.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>