[tools] perf: Left-align output contents
Message-id: <1418163608-7772-177-git-send-email-jolsa@redhat.com>
Patchwork-id: 101387
O-Subject: [PATCH RHEL7.1 BZ1169436 176/359] perf tools: Left-align output contents
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit 8246de88e95ddef7508f5601d7af85c3ab9e476b
Author: Namhyung Kim <namhyung@kernel.org>
Date: Thu Jul 31 14:47:35 2014 +0900

description
===========
Now perf left-aligns column headers but the contents does not.  It
should have same alignment.  This requires a change in pid sort key - it
consists of two part (pid and comm).  As length of comm can be vary it'd
be better to change the order of them.

Thanks to Jiri Olsa for pointing this out.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>