[tools] perf: Fix column alignment when headers aren't shown on TUI
Message-id: <1418163608-7772-183-git-send-email-jolsa@redhat.com>
Patchwork-id: 101393
O-Subject: [PATCH RHEL7.1 BZ1169436 182/359] perf tools: Fix column alignment when headers aren't shown on TUI
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit 59dc9f2534569d11a55c8b5dbe93c36f2b2fa506
Author: Namhyung Kim <namhyung@kernel.org>
Date: Thu Jul 31 14:47:41 2014 +0900

description
===========
If user sets ui.show-headers config option to false, it didn't calculate
default column width so it broke the alignment.  This is because it does
the calculation just before showing headers.

Move it to the beginning of the hist browser so that it can be called
regardless of the config option.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>