[tools] perf/evlist: Add 'system_wide' option
Message-id: <1418163608-7772-217-git-send-email-jolsa@redhat.com>
Patchwork-id: 101427
O-Subject: [PATCH RHEL7.1 BZ1169436 216/359] perf evlist: Add 'system_wide' option
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit bf8e8f4b832972c76d64ab2e2837a48397144887
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Thu Jul 31 09:00:51 2014 +0300

description
===========
Add an option to cause a selected event to be opened always without a
pid when configured by perf_evsel__config().

This is needed when using the sched_switch tracepoint to follow object
code execution.

sched_switch occurs before the task switch and so it cannot record it in
a context limited to that task.  Note that also means that sched_switch
is useless when capturing data per-thread, as is the 'context-switches'
software event for the same reason.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>