[tools] perf/probe: Warn user to rebuild target with debuginfo
Message-id: <1418163608-7772-226-git-send-email-jolsa@redhat.com>
Patchwork-id: 101436
O-Subject: [PATCH RHEL7.1 BZ1169436 225/359] perf probe: Warn user to rebuild target with debuginfo
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit 92561cb7883194714475c7a7775a11a9c40f75cb
Author: Masami Hiramatsu <masami.hiramatsu.pt@hitachi.com>
Date: Fri Aug 15 01:44:32 2014 +0000

description
===========
Warn user to rebuild target with debuginfo when the perf probe fails to
find debug information in the target binary.

Without this, perf probe just reports the failure, but it's no hint for
users. This gives more hint for users.

Without this:

  $ strip perf
  $ ./perf probe -x perf -L argv_split
  Failed to open debuginfo file.
    Error: Failed to show lines.

With this:

  $ strip perf
  $ ./perf probe -x perf -L argv_split
  The /home/fedora/ksrc/linux-3/tools/perf/perf file has no debug information.
  Rebuild with -g, or install an appropriate debuginfo package.
    Error: Failed to show lines.

The "rebuild with ..." part changes to "rebuild with CONFIG_DEBUG_INFO"
if the target is the kernel or a kernel module.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>