[tools] perf: Let a user specify a PMU event without any config terms
Message-id: <1418163608-7772-263-git-send-email-jolsa@redhat.com>
Patchwork-id: 101473
O-Subject: [PATCH RHEL7.1 BZ1169436 262/359] perf tools: Let a user specify a PMU event without any config terms
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit ad96227349901838e1a7f96f1dc22d96a97520c0
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Fri Aug 15 22:08:40 2014 +0300

description
===========
This enables a PMU event to be specified in the form:

	pmu//

which is effectively the same as:

	pmu/config=0/

This patch is a precursor to defining default config for a PMU.

Further explanation extracted from lkml thread:

Imagine that the 'tsc' term did not exist.

Intel PT trace data would not contain TSC packets, and the decoder would
not know how to decode them.

Then imagine that a new version of the hardware adds 'tsc'.

It is such a useful feature that we want it by default, but older
versions of the tools don't know how to decode it, so the kernel cannot
turn it on by default.

It is similar to why the kernel does not select perf_event_attr.mmap2 by
default.

The kernel doesn't know whether the tool supports it.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>