[tools] perf: Let default config be defined for a PMU
Message-id: <1418163608-7772-265-git-send-email-jolsa@redhat.com>
Patchwork-id: 101475
O-Subject: [PATCH RHEL7.1 BZ1169436 264/359] perf tools: Let default config be defined for a PMU
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit dc0a6202421170a6d8d2c6f5176575b3f60e0f85
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Thu Jul 31 09:00:49 2014 +0300

description
===========
This allows default config terms to be provided for a PMU. So, for
example, when the Intel PT PMU is added, it will be possible to specify:

	intel_pt//

which will be the same as:

	intel_pt/tsc=1,noretcomp=0/

meaning that the trace should contain TSC timestamps and perform 'return
compression'.

An important consideration of this patch is that it must be possible to
overwrite the default values.  That has meant changing the logic so that
a zero value can replace a non-zero value.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>