[tools] perf/evlist: Introduce perf_evlist__filter_pollfd method
Message-id: <1418163608-7772-276-git-send-email-jolsa@redhat.com>
Patchwork-id: 101486
O-Subject: [PATCH RHEL7.1 BZ1169436 275/359] perf evlist: Introduce perf_evlist__filter_pollfd method
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit 1ddec7f0d0ab5b71cf2cc5a782441c20e7afbcfb
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Aug 12 23:04:11 2014 -0300

description
===========
To remove all entries in evlist->pollfd[] that have revents matching at
least one of the bits in the specified mask.

It'll adjust evlist->nr_fds to the number of unfiltered fds and will
return this value, as a convenience and to avoid requiring direct access
to internal state of perf_evlist objects.

This will be used after polling the evlist fds so that we remove fds
that were closed by the kernel.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>