[tools] perf/evlist: Monitor POLLERR and POLLHUP events too
Message-id: <1418163608-7772-278-git-send-email-jolsa@redhat.com>
Patchwork-id: 101488
O-Subject: [PATCH RHEL7.1 BZ1169436 277/359] perf evlist: Monitor POLLERR and POLLHUP events too
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit 8179672c2f7b9c41a7ef3e8c907d214fa92ed614
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Wed Aug 13 11:26:21 2014 -0300

description
===========
We want to know when the fd went away, like when a monitored thread
exits.

If we do not monitor such events, then the tools will wait forever on
events from a vanished thread, like when running:

 $ sleep 5s &
 $ perf record -p `pidof sleep`

This builds upon the kernel patch by Jiri Olsa that actually makes a
poll on those file descriptors to return POLLHUP.

It is also needed to change the tools to use
perf_evlist__filter_pollfd() to check if there are remainings fds to
monitor or if all are gone, in which case they will exit the
poll/mmap/read loop.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>