[tools] perf/trace: Fix mmap return address truncation to 32-bit
Message-id: <1418163608-7772-301-git-send-email-jolsa@redhat.com>
Patchwork-id: 101511
O-Subject: [PATCH RHEL7.1 BZ1169436 300/359] perf trace: Fix mmap return address truncation to 32-bit
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit 2c82c3ad56921c47f28af9eb8ed96b6d99b47623
Author: Chang Hyun Park <heartinpiece@gmail.com>
Date: Fri Sep 26 21:54:01 2014 +0900

description
===========
Using 'perf trace' for mmap is truncating return values by stripping the
top 32 bits, actually printing only the lower 32 bits.

This was because the ret value was of an 'int' type and not a 'long'
type.

  The Problem:

  991258501.244 ( 0.004 ms): mmap(len: 40001536, prot: READ|WRITE, flags: PRIVATE|ANONYMOUS, fd: -1) = 0x56691000
  991258501.257 ( 0.000 ms): minfault [_int_malloc+0x1038] => //anon@0x7fa056691008 //(d.)

The first line shows an mmap, which succeeds and returns 0x56691000.

However the next line shows a memory access to that virtual memory area,
specifically to 0x7fa056691008. The upper 32 bit is lost due to the
problem mentioned above, and thus mmap's return value didn't have the
upper 0x7fa0.

Tested on 3.17-rc5 from the linus's tree, and the HEAD of tip/master
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>