[tools] perf/symbols: Encapsulate dsos list head into struct dsos
Message-id: <1418163608-7772-304-git-send-email-jolsa@redhat.com>
Patchwork-id: 101514
O-Subject: [PATCH RHEL7.1 BZ1169436 303/359] perf symbols: Encapsulate dsos list head into struct dsos
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit 8fa7d87f91479f7124142ca4ad93a37b80f8c1c0
Author: Waiman Long <Waiman.Long@hp.com>
Date: Mon Sep 29 16:07:28 2014 -0400

description
===========
This is a precursor patch to enable long name searching of DSOs using
a rbtree.

In this patch, a new dsos structure is created which contains only a
list head structure for the moment.

The new dsos structure is used, in turn, in the machine structure for
the user_dsos and kernel_dsos fields.

Only the following 3 dsos functions are modified to accept the new dsos
structure parameter instead of list_head:

 - dsos__add()
 - dsos__find()
 - __dsos__findnew()

[ Move struct dsos to dso.h to reduce the dso methods depends on machine.h ]
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>