[tools] perf/machine: Add missing dsos->root rbtree root initialization
Message-id: <1418163608-7772-327-git-send-email-jolsa@redhat.com>
Patchwork-id: 101536
O-Subject: [PATCH RHEL7.1 BZ1169436 326/359] perf machine: Add missing dsos->root rbtree root initialization
Bugzilla: 1169436
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

Bugzilla: 1169436
https://bugzilla.redhat.com/show_bug.cgi?id=1169436

upstream
========
commit e167f995e26249aa93708589c5eea539652351fa
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Oct 14 15:07:48 2014 -0300

description
===========
A segfault happens on 'perf test hists_link' because we end up using a
struct machines on the stack, and then machines__init() was not
initializing the newly introduced rb_root, just the existing list_head.

When we introduced struct dsos, to group the two ways to store dsos,
i.e. the linked list and the rbtree, we didn't turned the initialization
done in:

	machines__init(machines->host) ->
		machine__init() ->
			INIT_LIST_HEAD

into a dsos__init() to keep on initializing the list_head but _as well_
initializing the rb_root, oops.

All worked because outside perf-test we probably zalloc the whole thing
which ends up initializing it in to NULL.

So the problem looks contained to 'perf test' that uses it on stack,
etc.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>